refactor: remove unused WebP images and update video asset paths to use hyphenated naming
